requirements.txt
PyVESC
pyserial
